+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Make `EventSourcery::Postgres::OptimisedEventPollWaiter#shutdown` private 
 - Updated `EventSourcery::Postgres::OptimisedEventPollWaiter#poll` to ensure that `#shutdown!` is run when an error is raised
 or when the loop stops
+- Remove dynamic emit events methods
 
 ### Added
 - Configure projector tracker table name via `EventSourcery::Postgres.configure`